+++ src/connection.py
@@ -155,6 +155,13 @@ class Connection(object):
 
     def disconnect(self):
         if self.connected:
+            # Close file descriptors before socket closing.
+            if self.fin:
+                self.fin.close()
+
+            if self.fout:
+                self.fout.close()
+
             # There can be errors on shutdown.
             # We need to close socket and reset state even if it happens.
             try:
@@ -163,6 +170,10 @@ class Connection(object):
             except socket.error as e:
                 logger.warning('Error on socket shutdown: %s', e)
 
+            self.socket.close()
+
+        # Socket can be constructed but not connected.
+        elif self.socket:
             self.socket.close()
 
         self.reset_state()